requirements.txt
fastapi>=0.116.0
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
python-multipart==0.0.18
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
email-validator==2.2.0
python-dotenv==1.0.1
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic[email]==2.10.5
psycopg2-binary>=2.9.0
google-genai>=1.0.0